Move logout handler to namespace and handle redirect target
